Add support for Solaris peer credentials.
This refactors the test code for linux somewhat.  Also we added the
Solaris Zone ID, which matches what we do for NNG.  This works for
both Solaris and illumos targets, but only when cgo is enabled.
